created 3 components for the three room categories
