+++ book-a-room-app/src/components/room-categories/group-rooms/GroupRooms.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+interface GroupRoomsProps {
+  onSelectCategory: () => void;
+}
+
+const GroupRooms: React.FC<GroupRoomsProps> = ({ onSelectCategory }) => {
+  return (
+    <button onClick={onSelectCategory} className="category-button">
+      Group Rooms
+    </button>
+  );
+};
+
+export default GroupRooms;
