+++ book-a-room-app/src/components/room-categories/hostels/Hostels.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+interface HostelsProps {
+  onSelectCategory: () => void;
+}
+
+const Hostels: React.FC<HostelsProps> = ({ onSelectCategory }) => {
+  return (
+    <button onClick={onSelectCategory} className="category-button">
+      Ensuite Bedrooms
+    </button>
+  );
+};
+
+export default Hostels;
